moved mixing and ventilation to wind.py
