+++ tests/models/abiotic/test_energy_balance.py
@@ -326,139 +326,6 @@ def test_update_air_temperature():
     )
 
 
-def test_calculate_ventilation_rate_scalar():
-    """Test calculate ventilation rate scalar."""
-
-    from virtual_ecosystem.models.abiotic.energy_balance import (
-        calculate_ventilation_rate,
-    )
-
-    ra = 50.0
-    h = 20.0
-    expected = 1.0 / 1000.0
-
-    result = calculate_ventilation_rate(ra, h)
-    assert np.isclose(result, expected)
-
-
-def test_calculate_ventilation_rate_array():
-    """Test calculate ventilation rate array."""
-
-    from virtual_ecosystem.models.abiotic.energy_balance import (
-        calculate_ventilation_rate,
-    )
-
-    ra = np.array([10.0, 50.0, 0.0])
-    h = np.array([2.0, 20.0, 1.0])
-    expected = np.array([5.0e-02, 1.0e-03, 1.0e06])
-
-    result = calculate_ventilation_rate(ra, h)
-    np.testing.assert_allclose(result, expected)
-
-
-def test_calculate_ventilation_rate_zero_denominator():
-    """Test calculate ventilation rate scalar."""
-
-    from virtual_ecosystem.models.abiotic.energy_balance import (
-        calculate_ventilation_rate,
-    )
-
-    ra = 0.0
-    h = 0.0
-    expected = 1.0 / 1e-6
-
-    result = calculate_ventilation_rate(ra, h)
-    assert np.isclose(result, expected)
-
-
-def test_calculate_mixing_coefficients():
-    """Test mixing coefficients."""
-    from virtual_ecosystem.models.abiotic.energy_balance import (
-        calculate_mixing_coefficients_canopy,
-    )
-
-    layer_midpoints = np.array([[0.1, 0.5, 0.9], [0.2, 0.6, 0.8]])
-    canopy_height = np.repeat(1.0, 3)
-    friction_velocity = np.repeat(0.3, 3)
-    k = 0.4
-    expected = np.array([[0.00972, 0.015, 0.00108], [0.01536, 0.01152, 0.00384]])
-
-    result = calculate_mixing_coefficients_canopy(
-        layer_midpoints, canopy_height, friction_velocity, k
-    )
-
-    assert result.shape == layer_midpoints.shape
-    assert np.all(result >= 0)
-    np.testing.assert_allclose(result, expected, rtol=1e-6)
-
-
-def test_mix_and_ventilate(dummy_climate_data, fixture_core_components):
-    """Test mixing and ventilation."""
-
-    from virtual_ecosystem.models.abiotic.energy_balance import (
-        mix_and_ventilate,
-    )
-
-    lystr = fixture_core_components.layer_structure
-    data = dummy_climate_data
-    input_variable = data["relative_humidity"][lystr.index_filled_atmosphere].to_numpy()
-
-    layer_thickness = np.array(
-        [
-            [2.0, 2.0, 2.0, 2.0],
-            [10.0, 10.0, 10.0, 10.0],
-            [5.0, 5.0, 5.0, 5.0],
-            [2.0, 2.0, 2.0, 2.0],
-            [0.25, 0.25, 0.25, 0.25],
-        ]
-    )
-    exp_result = np.array(
-        [
-            [77.700816, 65.401632, 53.102448, 40.803264],
-            [90.341644, 90.341644, 90.341644, 90.341644],
-            [92.70733, 92.70733, 92.70733, 92.70733],
-            [96.313381, 96.313381, 96.313381, 96.313381],
-            [100.0, 100.0, 100.0, 100.0],
-        ]
-    )
-
-    result = mix_and_ventilate(
-        input_variable=input_variable,
-        layer_thickness=layer_thickness,
-        mixing_coefficient=np.full((5, 4), 0.001),
-        ventilation_rate=np.array([0.01, 0.02, 0.03, 0.04]),
-        time_interval=3600.0,
-    )
-    np.testing.assert_allclose(result, exp_result)
-
-
-def test_advect_from_toplayer():
-    """Test advection of moisture from top layer."""
-
-    from virtual_ecosystem.models.abiotic.energy_balance import (
-        advect_water_from_toplayer,
-    )
-
-    specific_humidity = np.array([0.010, 0.008, 0.006])
-    layer_thickness = np.array([10.0, 10.0, 10.0])
-    density_air = np.array([1.2, 1.1, 1.0])
-    wind_speed = np.array([1.0, 0.5, 2.0])
-    time_interval = 3600.0
-
-    expected_specific_humidity = np.array([0, 0, 0])
-
-    result = advect_water_from_toplayer(
-        specific_humidity=specific_humidity,
-        layer_thickness=layer_thickness,
-        density_air=density_air,
-        wind_speed=wind_speed,
-        characteristic_length=100,
-        time_interval=time_interval,
-    )
-
-    np.testing.assert_allclose(result, expected_specific_humidity)
-
-
 def test_update_humidity_vpd():
     """Test update atmospheric humidity."""
