+++ virtual_ecosystem/models/abiotic/microclimate.py
@@ -178,14 +178,14 @@ def run_microclimate(
     aerodynamic_resistance_soil = data["aerodynamic_resistance_surface"].to_numpy()
 
     # Turbulent mixing coefficient, [m2 s-1] and ventilation rate [s-1] above canopy
-    mixing_coefficient = energy_balance.calculate_mixing_coefficients_canopy(
+    mixing_coefficient = wind.calculate_mixing_coefficients_canopy(
         layer_midpoints=layer_midpoints,
         canopy_height=canopy_height,
         friction_velocity=data["friction_velocity"].to_numpy(),
         von_karman_constant=core_constants.von_karmans_constant,
     )
 
-    ventilation_rate = energy_balance.calculate_ventilation_rate(
+    ventilation_rate = wind.calculate_ventilation_rate(
         aerodynamic_resistance=aerodynamic_resistance_canopy[0],
         characteristic_height=canopy_height,
     )
@@ -353,7 +353,7 @@ def run_microclimate(
         all_air_temperature[1 : len(canopy_temperature) + 1] = air_temperature_canopy
         all_air_temperature[-1] = surface_air_temperature
 
-        all_air_temperature = energy_balance.mix_and_ventilate(
+        all_air_temperature = wind.mix_and_ventilate(
             input_variable=all_air_temperature,
             layer_thickness=above_ground_layer_thickness,
             ventilation_rate=ventilation_rate,
